Fix infinite loop in parseFrontmatter and deduplicate stripLeadingHtmlComments
- Fix bug b1a6331d: Increment index when HTML comment has trailing content to prevent infinite loop
- Fix bug 036ea055: Extract stripLeadingHtmlComments to utils/text.ts and remove duplicates from generated-skill-validation.ts and crystallization-proposer.ts
